pyproject.toml
[project]
name = "building-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "dspy>=3.0.3",
    "mlflow>=3.4.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "ipython>=9.5.0",
]
